+++ protohaven_api/integrations/sheets.py
@@ -103,3 +103,31 @@ def get_sign_ins_between(start, end):
         if start <= t <= end:
             data["timestamp"] = t
             yield data
+
+def get_ops_budget_state():
+    """Returns ops budgeting state from shop manager logbook"""
+    sheet_id = get_config("sheets/shop_manager_logbook")
+    headers = [h.strip().lower() for h in get_sheet_range(sheet_id, "Budget Summary!A2:A")]
+    values = get_sheet_range(sheet_id, "Budget Summary!B2:B")
+    data = dict(zip(headers, values))
+    return data
+
+def get_ops_event_log(start=None, end=None):
+    """Returns all events logged in the shop manager logbook between start and end dates."""
+    sheet_id = get_config("sheets/shop_manager_logbook")
+    headers = get_sheet_range(sheet_id, "Event Log!A1:E")
+    for row in get_sheet_range(sheet_id, "Event Log!A2:E"):
+        data = dict(zip(headers, row))
+        t = safe_parse_datetime(data["Date"])
+        if (not start or start <= t) and (not end or t <= end):
+            data["timestamp"] = t
+            yield data
+
+def get_ops_inventory():
+    """Returns all inventory information in the shop manager logbook"""
+    sheet_id = get_config("sheets/shop_manager_logbook")
+    headers = get_sheet_range(sheet_id, "Inventory!A2:F")
+    last_update = dateparser.parse(get_sheet(_range(sheet_id, "Inventory!B1"))[0][0])
+    for row in get_sheet_range(sheet_id, "Inventory!A3:F"):
+        data.append(dict(zip(headers, row)))
+    return last_update, data
